+++ scanpipe/tests/__init__.py
@@ -89,6 +89,20 @@ package_data1 = {
     "package_uid": "pkg:deb/debian/adduser@3.118?uuid=610bed29-ce39-40e7-92d6-fd8b",
 }
 
+for_package_uid = "pkg:pypi/daglib@0.3.2?uuid=4d1f048b-a155-4f95-8cf9-185ab872ab4c"
+dependency_data1 = {
+    "purl": "pkg:pypi/dask",
+    "extracted_requirement": "dask<2023.0.0,>=2022.6.0",
+    "scope": "install",
+    "is_runtime": True,
+    "is_optional": False,
+    "is_resolved": False,
+    "dependency_uid": "pkg:pypi/dask?uuid=e656b571-7d3f-46d1-b95b-8f037aef9692",
+    "for_package_uid": for_package_uid,
+    "datafile_path": "daglib-0.3.2.tar.gz-extract/daglib-0.3.2/PKG-INFO",
+    "datasource_id": "pypi_sdist_pkginfo",
+}
+
 license_policies = [
     {
         "license_key": "apache-2.0",