+++ scanpipe/views.py
@@ -47,6 +47,8 @@ from django_filters.views import FilterView
 
 from scancodeio.auth import ConditionalLoginRequired
 from scancodeio.auth import conditional_login_required
+from scanpipe.api.serializers import DiscoveredDependencySerializer
+from scanpipe.filters import DependencyFilterSet
 from scanpipe.filters import ErrorFilterSet
 from scanpipe.filters import PackageFilterSet
 from scanpipe.filters import ProjectFilterSet
@@ -56,6 +58,7 @@ from scanpipe.forms import AddPipelineForm
 from scanpipe.forms import ArchiveProjectForm
 from scanpipe.forms import ProjectForm
 from scanpipe.models import CodebaseResource
+from scanpipe.models import DiscoveredDependency
 from scanpipe.models import DiscoveredPackage
 from scanpipe.models import Project
 from scanpipe.models import ProjectError
@@ -554,6 +557,11 @@ class ProjectDetailView(ConditionalLoginRequired, ProjectViewMixin, generic.Deta
             "type",
             "license_expression",
         )
+        dependencies = project.discovereddependencys.all().only(
+            "is_runtime",
+            "is_optional",
+            "is_resolved",
+        )
 
         file_languages = files.values_list("programming_language", flat=True)
         file_mime_types = files.values_list("mime_type", flat=True)
@@ -570,6 +578,10 @@ class ProjectDetailView(ConditionalLoginRequired, ProjectViewMixin, generic.Deta
 
         package_licenses = packages.values_list("license_expression", flat=True)
         package_types = packages.values_list("type", flat=True)
+
+        dependency_is_runtime = dependencies.values_list("is_runtime", flat=True)
+        dependency_is_optional = dependencies.values_list("is_optional", flat=True)
+        dependency_is_resolved = dependencies.values_list("is_resolved", flat=True)
 
         inputs, missing_inputs = project.inputs_with_source
         if missing_inputs:
@@ -606,6 +618,9 @@ class ProjectDetailView(ConditionalLoginRequired, ProjectViewMixin, generic.Deta
                 "file_compliance_alert": self.get_summary(file_compliance_alert),
                 "package_licenses": self.get_summary(package_licenses),
                 "package_types": self.get_summary(package_types),
+                "dependency_is_runtime": self.get_summary(dependency_is_runtime),
+                "dependency_is_optional": self.get_summary(dependency_is_optional),
+                "dependency_is_resolved": self.get_summary(dependency_is_resolved),
                 "file_filter": file_filter,
                 "add_pipeline_form": AddPipelineForm(),
                 "add_inputs_form": AddInputsForm(),
@@ -868,6 +883,18 @@ class DiscoveredPackageListView(
     ]
 
 
+class DiscoveredDependencyListView(
+    ConditionalLoginRequired,
+    ProjectRelatedViewMixin,
+    PaginatedFilterView,
+):
+    model = DiscoveredDependency
+    filterset_class = DependencyFilterSet
+    template_name = "scanpipe/dependency_list.html"
+    paginate_by = 100
+    prefetch_related = ["discovered_packages"]
+
+
 class ProjectErrorListView(
     ConditionalLoginRequired,
     ProjectRelatedViewMixin,
@@ -1026,6 +1053,18 @@ class DiscoveredPackageDetailsView(
             "icon_class": "fas fa-database",
         },
     }
+
+
+class DiscoveredDependencyDetailsView(
+    ConditionalLoginRequired, ProjectRelatedViewMixin, generic.DetailView
+):
+    model = DiscoveredDependency
+    template_name = "scanpipe/dependency_detail.html"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["dependency_data"] = DiscoveredDependencySerializer(self.object).data
+        return context
 
 
 @conditional_login_required